Apply Trolltech Qt4 patch
git-svn-id: svn://scribus.net/branches/Version13x-Qt4/Scribus@3856 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/prefstable.cpp
@@ -56,7 +56,7 @@ QString PrefsTable::get(int row, int col, const QString& defValue)
 {
 	checkSize(row, col, defValue);
 	if ((*table[row])[col] == "__NOT__SET__")
-		table[row]->insert(table[row]->at(col), defValue);
+		table[row]->insert(table[row]->at(col).toInt(), defValue);
 
 	return (*table[row])[col];
 }
@@ -74,7 +74,7 @@ void PrefsTable::set(int row, int col, const std::string& value)
 void PrefsTable::set(int row, int col, const QString& value)
 {
 	checkSize(row, col, "__NOT__SET__");
-	table[row]->insert(table[row]->at(col), value);
+	table[row]->insert(table[row]->at(col).toInt(), value);
 }
 
 int PrefsTable::getInt(int row, int col, int defValue)
